command allowance enforcement implemented and tested

+++ usercommandmgmt/commanddatahandler.py
@@ -14,6 +14,9 @@ class Database:
 
 
     def ReadFromDB(self):
+        '''
+        reads from the database
+        '''
         try:
             with open(self.dir, "r") as f:
                 self.loaded_cmd_data = json.load(f)
@@ -24,11 +27,11 @@ class Database:
         '''
         saves a command's info to the database
         '''
-        n_entry = {"cmd_name": cmd_name,
+        n_entry = {cmd_name : [{
                    "cmd_owner": cmd_owner,
-                   "admin-made": is_admin}
+                   "admin-made": is_admin
+        }]}
         self.loaded_cmd_data["db"].append(n_entry)
-        print(self.loaded_cmd_data)
         with open(self.dir, 'w') as outfile:
             json.dump(self.loaded_cmd_data, outfile)
 
@@ -36,7 +39,16 @@ class Database:
         '''
         removes an entry from the list of command data and updates the database
         '''
-        del self.loaded_cmd_data[cmd_name]
+        if not self.CommExists(cmd_name):
+            return
+
+        remove_index = 0
+        for cmd in self.loaded_cmd_data["db"]:
+            if list(cmd.items())[0][0] == cmd_name:
+                break
+            remove_index += 1
+            
+        self.loaded_cmd_data["db"].pop(remove_index)
         with open(self.dir, 'w') as outfile:
             json.dump(self.loaded_cmd_data, outfile)
 
@@ -45,9 +57,8 @@ class Database:
         returns how many commands are owned by the user who owns the given ID. If user does not exist in database, returns zero.
         """
         cmd_count = 0
-        print(self.loaded_cmd_data)
         for command in self.loaded_cmd_data["db"]:
-            if command["cmd_owner"] == user_id:
+            if list(command.items())[0][1][0]['cmd_owner'] == user_id and list(command.items())[0][1][0]["admin-made"] == False:
                 cmd_count += 1
 
         return cmd_count
@@ -57,7 +68,7 @@ class Database:
         returns true if the given command name exists in the database
         """
         for cmd in self.loaded_cmd_data["db"]:
-            if cmd["cmd_name"] == cmd_name:
+            if list(cmd.items())[0][0] == cmd_name:
                 return True
 
         return False
@@ -67,7 +78,7 @@ class Database:
         returns the database entry for the given cmd_name, if it exists
         """
         for cmd in self.loaded_cmd_data["db"]:
-            if cmd["cmd_name"] == cmd_name:
+            if list(cmd.items())[0][0] == cmd_name:
                 return cmd
 
         return None
@@ -78,10 +89,10 @@ class Database:
         return false if the command does not exist.
         """
 
-        if CommExists(cmd_name) == False:
+        if self.CommExists(cmd_name) == False:
             return False
 
-        return self.GetComm(cmd_name)["cmd_name"] == user_id
+        return list(self.GetComm(cmd_name).items())[0][1][0]['cmd_owner'] == user_id
